+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.0
 - Require Sidekiq 3.
+- Linear Scaling Strategy will not scale down past number of active workers. Assumes 1-1 SK process/dyno mapping.
 - QueueSystem#workers returns the number of engaged SK processes.
 
 ## 0.9.0